Update Day-3-logics
Find the Tiger Game
Welcome to the "Find the Tiger" game! This is a simple and fun text-based adventure game created with Python. Navigate through the jungle and find the tiger to win!

How to Play
Start the Game: Begin your jungle adventure.
Choose a Direction: Pick North, South, East, or West.
Find the Tiger: Make the right choice to find the tiger and win