Create Day-7
Hangman Game
Welcome to the Hangman game! In this classic word-guessing game, you will try to guess the letters of a randomly chosen word. You have 6 lives, and each incorrect guess will cost you one life. Can you guess the word before you run out of lives?

How to Play:
Start the Game: The game will choose a word for you to guess.
Make a Guess: Enter a letter you think is in the word.
Correct Guess: If the letter is in the word, it will be revealed in its correct position.
Incorrect Guess: If the letter is not in the word, you lose a life.
Win or Lose: The game ends when you guess the word correctly or lose all your lives.
Good luck and have fun!